fastapi/pyproject.toml
[tool.poetry]
name = "dodam"
version = "0.1.0"
description = ""
authors = ["kookim2jung <kookim2jung@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.4"
aiosignal = "1.3.1"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
attrs = "23.2.0"
authlib = "1.3.0"
bcrypt = "4.1.2"
boto3 = "1.34.88"
botocore = "1.34.88"
certifi = "^2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "42.0.5"
distro = "1.9.0"
dnspython = "2.6.1"
ecdsa = "0.19.0"
fastapi = "0.110.1"
frozenlist = "1.4.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
jmespath = "1.0.1"
libmagic = "1.0"
mako = "1.3.3"
markupsafe = "2.1.5"
motor = "3.4.0"
multidict = "6.0.5"
openai = "0.28.0"
pyasn1 = "0.6.0"
pycparser = "2.22"
pydantic-core = "2.18.1"
pyjwt = "2.8.0"
pymongo = "4.6.3"
pymysql = "1.1.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.9"
requests = "2.31.0"
rsa = "4.9"
s3transfer = "0.10.1"
security = "1.2.1"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.29"
starlette = "0.37.2"
tqdm = "4.66.2"
typing-extensions = "4.11.0"
urllib3 = "1.26.18"
uvicorn = "0.29.0"
yarl = "1.9.4"
pinecone-client = "^4.0.0"
numpy = "^1.26.4"
pydantic = "2.7.0"
python-magic = "^0.4.27"
#python-magic-bin = "^0.4.14"
responses = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
